Use giflib/boost.endian to read/write GIF files with fractint extensions
Because the maximum length of an extension data block is 255 bytes, it's
possible for a 16-bit int to be split across a data block boundary.
Check the function codes of extension blocks to distinguish between an
application extension block and a continuation data block.  Accumulate
all ranges bytes into a contiguous array before decoding the integers in
the data.