Make all typed input conistent in <>'s, lowercase filenames, drop DOS talk
Write <B> when the letter b can be typed and write <Shift+B> when the
uppercase form must be typed.  This is only in a few places such as the
image browser when Shift+D deletes a file and Shift+R renames a file.

Use Alt+, Ctrl+ and Shift+ to indicate modifiers instead of Alt-, Ctrl-
and Shift-.  The + helps reinforce that both keys should be held down
simultaneously.