Update palette map section; rename to Color Maps
Displays haven't been palette based for a long time; they're all
truecolor now.  The general process of turning an iteration into a color
is the process of "mapping an iteration to a color" (in the mathematical
sense of map), so renaming the section to Color Maps is more appropriate
now.