Drop over eager assert
The slide show code wants to display a message, but there's no way to
know if a text window or a graphics window is active, so it wants to
display the string in both modes simultaneously.  So it's OK to draw
a plot string while a text window is displayed.