Correct slide show message display
The slide show wants to display a message, so needs to know if we are
displaying text or graphics so it can make the appropriate call.
Really there should be a dedicated piece of screen real estate for the
slide show to do it's message display, but that will wait until we
are using wxWidgets.

Use std::this_thread::sleep_for to sleep in 10ms chunks while checking
for key presses.  Since waiting for a keypress is how the slide show
mechanism works, we have to turn off the playback mode while waiting and
then turn it back on, otherwise we won't delay for the requested amount
of time, we'll just get the next key stroke in the slide show.