Replace Macros With Const References to g_params
They can't be globals because they must maintain "live" references to
the g_params array elements.  They are simply aliased names for the
parameters at specific indices.

Another option would be to introduce an enum for the indices and write
g_params[LAMBDA], but this is a smaller code change.