Use different CMake presets for CI workflows
Workflows perform configure, build, test and package steps and use a
child of the source directory as the build directory.  This eliminates
the use of relative paths that actions/upload-artifact doesn't like.

The package step will run the install step, so we don't need the OS
specific names of the install or package targets.