Drop surrounding spaces and double quotes from link name
When searching for a link by name, the surrounding space/quotes were
ignored and the link name wasn't used for anything other than a lookup
previously.  Now the AsciiDoc compiler is using the name to form markup
and the spaces/quotes just get in the way.  Rather than doing additional
processing when compiling AsciiDoc, just drop the surrounding characters
when constructing the name in the fist place.