Guard against registry API failing
When "id makedoc savedir=." is run, but no saved window position had
ever been written into the registry, an assert would occur in Debug
builds.

The problem was that the window position would be computed as
{CW_USEDEFAULT, CW_USEDEFAULT}, where CW_USEDEFAULT is 0x80000000.  Then
the code would attempt to see if this position was on screen, and it
wouldn't be, so then the code would attempt to delete a saved window
position that was off-screen, assuming that the last position was on a
monitor that is now detached.  However, since the registry key didn't
already exist, the code to forget the window position would fail.

Fix this by only looking to forget the window position if we actually
got one from the registry *and* it was off screen.

Also add more error checking around all registry API calls.

Fixes #147