Rename LDBL to LDouble
While x86 CPU implements an 80-bit extended precision double that adds 3
more decimal digits of precision compared to a 64-bit double, the
80-bit double isn't supported in the SIMD extensions, which only operate
on 32-bit floats and 64-bit doubles.  The 80-bit double also isn't
supported by MSVC in 64-bit compilation.

Most of the code in Id assumes that LDouble is natively supported by the
CPU as a primitive type and uses this type for variables throughout.