Encapsulate timer function variants
Really, it's three separate functions for different use cases that share
a common preamble and postamble.  The functions have different
arguments, which is probably why it was created as a varargs function,
but varargs functions are notoriously easy to misuse, so encapsulate
the varargs implementation behind three functions for the different use
cases.