Mark member variable used
The BITMAPINFO and RGBQUAD elements are used through a pointer to the
m_bmi member and must be allocated adjacent to each other in memory.
It's a bit of a hack to mke them adjacent member variables instead of
allocating a chunk of memory of the right size and using pointers.