Restore mouse support to orbits window
The old mouse code worked by faking up key presses for all the mouse
movements.  We want to move away from this style of mouse handling and
move towards handling mouse events directly.  The orbit window handling
would block waiting for keyboard input through the cursor.  Our indirect
method of handling mouse events updates the cursor position, so add a
flag on the Cursor object that unblocks waiting for a key press if the
position was updated externally.  This unblocks the Cursor key press
code once and allows the orbits window to continue.