Store Id version separately from legacy version in GIF files
Because Id reset to version 1.0 after forking from FRACTINT,
we need to store our version number separately from the 'release' field
used by FRACTINT in the GIF extension block.  Increase the version
number of the FractalInfo extension block and store the version as new
fields at the end of the block, replacing some of the 'future' array.

This means that GIF files written by Id might not be interpreted
correctly by FRACTINT anymore.

Store the version read from the file as g_file_version; this value is
used to adjust the interpretation of any GIF extension block values when
a file is read.  Restore some version-checked backward compatibility code
from FRACTINT that adjusts variables based on the data read from the file.

Id dropped backwards compatibility with old versions of FRACTINT before
the version of the fork to simplify the code going forward.  This may
result in some differences in images between FRACTINT and Id when the
old behavior is activated in FRACTINT.

The reset parameter sets g_version that is used for any backward
compatibility checks needed going forward.

Fixes #240