Use Driver::get_filename to prompt the user for a filename
On Windows, this invokes the common dialog GetOpenFileName.
On Unix, this is stubbed out.

Both will be replaced by the corresponding wxWidgets common dialog in
version 2.0.

This eliminates a whole bunch of special-purpose logic for browsing
directories, collecting filenames, and so-on as a full-screen choice
screen.

Move CurrentPathSaver.h from tests to libid/io.

Rename locate_config_file to locate_input_file and use it when
reading GIF files from the command-line.

Fixes #258
Fixes #267