Restore handling of commands as command-line arguments
The assumption that an argument couldn't be found as an input image file
should terminate with stop_msg was wrong.  First, a check is made to see
if the argument is a file that can be opened and has a GIF signature and
if this fails, then the argument is processed as a parameter command.

By unconditionally assuming failure when the file is not located,
parameter command arguments aren't executed properly.

Fixes #269