Fix variable corruption in tab_display
There was a parameter prompt that exceeded the char array previously
declared of size 50.  However, the prompt string returned is from the
FractalSpecific table where all the strings are present as constant
data.  So instead of accepting a pointer to a buffer and copying the
parameter prompt into the buffer, return the pointer to the parameter
prompt directly from the FractalSpecific structure.

Additionally, there were two use cases of type_has_param:
1. Just know if there is a parameter for the given param number
2. ...and additionally return the prompt string

Add an overload to distinguish these two cases.