Implement WxDriver::schedule_alarm
Is this doing anything anymore?  It's only called from one place in
framain2.cpp and in the Win32 implementation, all that happens is that a
timer is setup and killed when the screen is redrawn.  The timer event
itself is dropped on the floor, so when the timer fires, nothing special
happens.  Therefore, it doesn't impact behavior.