Enable rotate image test
The FRACTINT gold and the rendered image differ by 3.8% and a threshold
of 3% is hard-coded into the image compare tool.  So just update the
gold image from Id and enable the test.  This will still catch large
regressions.  Another option would be to enhance the image compare tool
to be configurable with a threshold on a per-test basis, but this seems
simpler.