Ignore top-level id.html, id.pdf files
If the AsciiDoc tools for processing id.adoc are not on the local
system, but the built files exist in the top-level directory, then they
are installed from there.  This lets you test the full install logic
locally by downloadin the HTML and PDF help files from a github workflow
run.

As a consequence of all that, the files shouldn't be added to source
control at the top-level, so ignore them.