Separate responsibilities of coloring method vs. color index
Magic negative numbers were used for inside and outside color indices to
indicate coloring methods that weren't based on an index to the
colormap.  In order to extend the coloring methods and in order to
extend colors beyond simple indices, these responsibilities need to be
separated.

Backwards compatibility with image and parameter files is maintained.