Use a single job to update github pages
With multiple jobs running concurrently, there's a race condition on
which job completes first getting the commit.  So collect all the static
analysis results together in a single job and use that to push to github
pages.  Unfortunately this means that the results are updated after the
longest pole of all the analysis jobs.